Create package for mappable items
