+++ collect_app/src/test/java/org/odk/collect/android/widgets/items/SelectOneFromMapDialogFragmentTest.kt
@@ -38,10 +38,11 @@ import org.odk.collect.android.widgets.utilities.WidgetAnswerDialogFragment.Comp
 import org.odk.collect.androidshared.ui.FragmentFactoryBuilder
 import org.odk.collect.async.Scheduler
 import org.odk.collect.fragmentstest.FragmentScenarioLauncherRule
-import org.odk.collect.geo.selection.IconifiedText
-import org.odk.collect.geo.selection.MappableItem
+import org.odk.collect.geo.items.IconifiedText
+import org.odk.collect.geo.items.MappableItem
 import org.odk.collect.geo.selection.SelectionMapFragment
 import org.odk.collect.geo.selection.SelectionMapFragment.Companion.REQUEST_SELECT_ITEM
+import org.odk.collect.icons.R
 import org.odk.collect.maps.MapFragment
 import org.odk.collect.maps.MapFragmentFactory
 import org.odk.collect.maps.MapPoint
@@ -194,10 +195,10 @@ class SelectOneFromMapDialogFragmentTest {
                                 firstFeatureGeometry[2].toDouble(),
                                 firstFeatureGeometry[3].toDouble()
                             ),
-                            smallIcon = org.odk.collect.icons.R.drawable.ic_map_marker_with_hole_small,
-                            largeIcon = org.odk.collect.icons.R.drawable.ic_map_marker_with_hole_big,
+                            smallIcon = R.drawable.ic_map_marker_with_hole_small,
+                            largeIcon = R.drawable.ic_map_marker_with_hole_big,
                             action = IconifiedText(
-                                org.odk.collect.icons.R.drawable.ic_save,
+                                R.drawable.ic_save,
                                 application.getString(org.odk.collect.strings.R.string.select_item)
                             )
                         ),
@@ -210,10 +211,10 @@ class SelectOneFromMapDialogFragmentTest {
                                 secondFeatureGeometry[2].toDouble(),
                                 secondFeatureGeometry[3].toDouble()
                             ),
-                            smallIcon = org.odk.collect.icons.R.drawable.ic_map_marker_with_hole_small,
-                            largeIcon = org.odk.collect.icons.R.drawable.ic_map_marker_with_hole_big,
+                            smallIcon = R.drawable.ic_map_marker_with_hole_small,
+                            largeIcon = R.drawable.ic_map_marker_with_hole_big,
                             action = IconifiedText(
-                                org.odk.collect.icons.R.drawable.ic_save,
+                                R.drawable.ic_save,
                                 application.getString(org.odk.collect.strings.R.string.select_item)
                             )
                         )
